pull network::constants into network module
