+++ examples/handshake.rs
@@ -6,10 +6,11 @@ use std::{env, process};
 use std::io::Write;
 
 use bitcoin::consensus::encode;
-use bitcoin::network::{address, constants, message};
+use bitcoin::network::{address, message};
 use bitcoin::network::stream_reader::StreamReader;
 use bitcoin::secp256k1;
 use bitcoin::secp256k1::rand::Rng;
+use bitcoin::{Network, ServiceFlags};
 
 fn main() {
     // This example establishes a connection to a Bitcoin node, sends the intial
@@ -30,7 +31,7 @@ fn main() {
     let version_message = build_version_message(address);
 
     let first_message = message::RawNetworkMessage {
-        magic: constants::Network::Bitcoin.magic(),
+        magic: Network::Bitcoin.magic(),
         payload: version_message,
     };
 
@@ -50,7 +51,7 @@ fn main() {
                     println!("Received version message: {:?}", reply.payload);
 
                     let second_message = message::RawNetworkMessage {
-                        magic: constants::Network::Bitcoin.magic(),
+                        magic: Network::Bitcoin.magic(),
                         payload: message::NetworkMessage::Verack,
                     };
 
@@ -78,7 +79,7 @@ fn build_version_message(address: SocketAddr) -> message::NetworkMessage {
     let my_address = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0)), 0);
 
     // "bitfield of features to be enabled for this connection"
-    let services = constants::ServiceFlags::NONE;
+    let services = ServiceFlags::NONE;
 
     // "standard UNIX timestamp in seconds"
     let timestamp = SystemTime::now()
@@ -87,10 +88,10 @@ fn build_version_message(address: SocketAddr) -> message::NetworkMessage {
         .as_secs();
 
     // "The network address of the node receiving this message"
-    let addr_recv = address::Address::new(&address, constants::ServiceFlags::NONE);
+    let addr_recv = address::Address::new(&address, ServiceFlags::NONE);
 
     // "The network address of the node emitting this message"
-    let addr_from = address::Address::new(&my_address, constants::ServiceFlags::NONE);
+    let addr_from = address::Address::new(&my_address, ServiceFlags::NONE);
 
     // "Node random nonce, randomly generated every time a version packet is sent. This nonce is used to detect connections to self."
     let nonce: u64 = secp256k1::rand::thread_rng().gen();
