+++ src/consensus/params.rs
@@ -17,8 +17,8 @@
 //! This module provides predefined set of parameters for different chains.
 //!
 
-use network::constants::Network;
 use util::uint::Uint256;
+use Network;
 
 /// Lowest possible difficulty for Mainnet. See comment on Params::pow_limit for more info.
 const MAX_BITS_BITCOIN: Uint256 = Uint256([
