+++ src/lib.rs
@@ -75,7 +75,8 @@ pub use primitives::transaction::TxOut;
 pub use primitives::transaction::OutPoint;
 pub use primitives::transaction::SigHashType;
 pub use consensus::encode::VarInt;
-pub use network::constants::Network;
+pub use network::Network;
+pub use network::ServiceFlags;
 pub use script::Script;
 pub use util::Error;
 pub use util::address::Address;
